services/sound-spleeter/requirements.txt
grpcio>=1.14.2
grpcio-tools>=1.14.1
requests==2.20.0
spleeter
